# backend/internal/accountservice/gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Accounts Backend Service","description":"Microservice for account management \u0026 operations","version":""},"host":"localhost:80","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/api/v1/account-service/accounts/login":{"post":{"tags":["account"],"summary":"login account","operationId":"account#login","parameters":[{"name":"LoginRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/AccountLoginRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/AccountLoginResponseBody","required":["email","token","refresh_token","role"]}}},"schemes":["http"]}},"/api/v1/account-service/accounts/register":{"post":{"tags":["account"],"summary":"register account","operationId":"account#register","parameters":[{"name":"RegisterRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/AccountRegisterRequestBody","required":["email","password","confirmedPassword","firstName","lastName"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/AccountRegisterResponseBody","required":["message"]}}},"schemes":["http"]}},"/api/v1/account-service/playlists":{"get":{"tags":["playlist"],"summary":"getAccountPlaylistCollection playlist","operationId":"playlist#getAccountPlaylistCollection","parameters":[{"name":"Authorization","in":"header","description":"JSON Web Token","required":false,"type":"string","pattern":"^Bearer [^ ]+$"},{"name":"GetAccountPlaylistCollectionRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/PlaylistGetAccountPlaylistCollectionRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/PlaylistGetAccountPlaylistCollectionResponseBody","required":["total","resources"]}}},"schemes":["http"]},"post":{"tags":["playlist"],"summary":"createAccountPlaylist playlist","operationId":"playlist#createAccountPlaylist","parameters":[{"name":"Authorization","in":"header","description":"JSON Web Token","required":false,"type":"string","pattern":"^Bearer [^ ]+$"},{"name":"CreateAccountPlaylistRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/PlaylistCreateAccountPlaylistRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/PlaylistCreateAccountPlaylistResponseBody","required":["message"]}}},"schemes":["http"]}},"/api/v1/account-service/playlists/{playlistID}":{"get":{"tags":["playlist"],"summary":"getAccountPlaylist playlist","operationId":"playlist#getAccountPlaylist","parameters":[{"name":"playlistID","in":"path","description":"Playlist ID","required":true,"type":"integer"},{"name":"Authorization","in":"header","description":"JSON Web Token","required":false,"type":"string","pattern":"^Bearer [^ ]+$"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/PlaylistGetAccountPlaylistResponseBody","required":["id","name","trackIDs"]}}},"schemes":["http"]},"put":{"tags":["playlist"],"summary":"renameAccountPlaylist playlist","operationId":"playlist#renameAccountPlaylist","parameters":[{"name":"playlistID","in":"path","description":"Playlist id to modify","required":true,"type":"integer"},{"name":"Authorization","in":"header","description":"JSON Web Token","required":false,"type":"string","pattern":"^Bearer [^ ]+$"},{"name":"RenameAccountPlaylistRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/PlaylistRenameAccountPlaylistRequestBody"}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/PlaylistRenameAccountPlaylistResponseBody","required":["message"]}}},"schemes":["http"]},"delete":{"tags":["playlist"],"summary":"deleteAccountPlaylist playlist","operationId":"playlist#deleteAccountPlaylist","parameters":[{"name":"playlistID","in":"path","required":true,"type":"integer"},{"name":"Authorization","in":"header","description":"JSON Web Token","required":false,"type":"string","pattern":"^Bearer [^ ]+$"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/PlaylistDeleteAccountPlaylistResponseBody","required":["message"]}}},"schemes":["http"]}},"/api/v1/account-service/playlists/{playlistID}/tracks/{trackID}":{"post":{"tags":["playlist"],"summary":"addTrackToAccountPlaylist playlist","operationId":"playlist#addTrackToAccountPlaylist","parameters":[{"name":"playlistID","in":"path","description":"Playlist ID to modify","required":true,"type":"integer"},{"name":"trackID","in":"path","description":"Track ID to be added","required":true,"type":"integer"},{"name":"Authorization","in":"header","description":"JSON Web Token","required":false,"type":"string","pattern":"^Bearer [^ ]+$"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/PlaylistAddTrackToAccountPlaylistResponseBody","required":["message"]}}},"schemes":["http"]},"delete":{"tags":["playlist"],"summary":"removeTrackFromAccountPlaylist playlist","operationId":"playlist#removeTrackFromAccountPlaylist","parameters":[{"name":"playlistID","in":"path","description":"Playlist ID to modify","required":true,"type":"integer"},{"name":"trackID","in":"path","description":"Track ID to be deleted","required":true,"type":"integer"},{"name":"Authorization","in":"header","description":"JSON Web Token","required":false,"type":"string","pattern":"^Bearer [^ ]+$"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/PlaylistRemoveTrackFromAccountPlaylistResponseBody","required":["message"]}}},"schemes":["http"]}}},"definitions":{"AccountLoginRequestBody":{"title":"AccountLoginRequestBody","type":"object","properties":{"email":{"type":"string","description":"Email of the user","example":"Vitae soluta sequi."},"password":{"type":"string","description":"Password of the user","example":"Architecto dolor ipsum repellat."}},"example":{"email":"Sed magnam.","password":"Fugiat quod tenetur facere reiciendis."}},"AccountLoginResponseBody":{"title":"AccountLoginResponseBody","type":"object","properties":{"accountID":{"type":"string","description":"User's role","example":"Ipsum suscipit."},"email":{"type":"string","description":"User's email","example":"Voluptas modi quo."},"refresh_token":{"type":"string","description":"Refresh token","example":"Omnis est debitis ut aliquid iste."},"role":{"type":"string","description":"User's role","example":"Deserunt similique aliquam consequatur ex ratione."},"token":{"type":"string","description":"JWT Token","example":"Qui ullam voluptas libero id est."}},"example":{"accountID":"Eos non rem reprehenderit mollitia et asperiores.","email":"Nobis asperiores voluptate saepe ut omnis voluptatem.","refresh_token":"Earum reprehenderit voluptatem rem sit voluptatum.","role":"Nulla velit totam excepturi.","token":"Est voluptate cumque et."},"required":["email","token","refresh_token","role"]},"AccountPlaylistResponseResponseBody":{"title":"AccountPlaylistResponseResponseBody","type":"object","properties":{"id":{"type":"integer","description":"Playlist id","example":1063571465,"format":"int32"},"name":{"type":"string","description":"Playlist name","example":"Suscipit culpa."},"trackIDs":{"type":"array","items":{"type":"integer","example":576478457,"format":"int32"},"description":"Array of TrackIDs","example":[1743088714,782442451,470044177]}},"example":{"id":1898964902,"name":"Et quo.","trackIDs":[1774012890,1320919651,1760146493]},"required":["id","name","trackIDs"]},"AccountRegisterRequestBody":{"title":"AccountRegisterRequestBody","type":"object","properties":{"confirmedPassword":{"type":"string","description":"Confirmed password of the user","example":"Aspernatur omnis."},"email":{"type":"string","description":"Email of the user","example":"Delectus rerum veniam ipsam iste et."},"firstName":{"type":"string","description":"First name of the user","example":"Voluptatem nisi assumenda consequatur nam."},"lastName":{"type":"string","description":"Last name of the user","example":"Et vitae dolore temporibus dolore."},"password":{"type":"string","description":"Password of the user","example":"Maxime recusandae."}},"example":{"confirmedPassword":"Eaque voluptas.","email":"Et culpa doloremque sint ex unde sequi.","firstName":"Quia tempora a minima.","lastName":"Atque est error voluptas dolorem.","password":"Asperiores et eius ducimus."},"required":["email","password","confirmedPassword","firstName","lastName"]},"AccountRegisterResponseBody":{"title":"AccountRegisterResponseBody","type":"object","properties":{"message":{"type":"string","description":"Operation completion status","example":"Accusamus occaecati velit."}},"example":{"message":"Asperiores est voluptatibus voluptas ut rem."},"required":["message"]},"PlaylistAddTrackToAccountPlaylistResponseBody":{"title":"PlaylistAddTrackToAccountPlaylistResponseBody","type":"object","properties":{"message":{"type":"string","description":"Operation completion status","example":"Odit sed corrupti atque ut."}},"example":{"message":"Id et culpa exercitationem eveniet fuga."},"required":["message"]},"PlaylistCreateAccountPlaylistRequestBody":{"title":"PlaylistCreateAccountPlaylistRequestBody","type":"object","properties":{"name":{"type":"string","description":"Playlist name","example":"Sint hic consequatur non dolorem nihil nostrum."}},"example":{"name":"Repudiandae omnis autem aut omnis vitae."}},"PlaylistCreateAccountPlaylistResponseBody":{"title":"PlaylistCreateAccountPlaylistResponseBody","type":"object","properties":{"message":{"type":"string","description":"Operation completion status","example":"Sequi cupiditate est atque fuga."}},"example":{"message":"Fuga quia consectetur rerum."},"required":["message"]},"PlaylistDeleteAccountPlaylistResponseBody":{"title":"PlaylistDeleteAccountPlaylistResponseBody","type":"object","properties":{"message":{"type":"string","description":"Operation completion status","example":"Maiores illum debitis."}},"example":{"message":"Illo libero et."},"required":["message"]},"PlaylistGetAccountPlaylistCollectionRequestBody":{"title":"PlaylistGetAccountPlaylistCollectionRequestBody","type":"object","properties":{"accountID":{"type":"integer","example":13145682454046792702,"format":"int64"}},"example":{"accountID":17017542213295389341}},"PlaylistGetAccountPlaylistCollectionResponseBody":{"title":"PlaylistGetAccountPlaylistCollectionResponseBody","type":"object","properties":{"resources":{"type":"array","items":{"$ref":"#/definitions/AccountPlaylistResponseResponseBody"},"description":"Operation completion status","example":[{"id":1978075829,"name":"Ipsam quis aut.","trackIDs":[1480478238,1629895974,393565112,1968683977]},{"id":1978075829,"name":"Ipsam quis aut.","trackIDs":[1480478238,1629895974,393565112,1968683977]},{"id":1978075829,"name":"Ipsam quis aut.","trackIDs":[1480478238,1629895974,393565112,1968683977]},{"id":1978075829,"name":"Ipsam quis aut.","trackIDs":[1480478238,1629895974,393565112,1968683977]}]},"total":{"type":"integer","description":"Number of resources","example":731017624,"format":"int32"}},"example":{"resources":[{"id":1978075829,"name":"Ipsam quis aut.","trackIDs":[1480478238,1629895974,393565112,1968683977]},{"id":1978075829,"name":"Ipsam quis aut.","trackIDs":[1480478238,1629895974,393565112,1968683977]},{"id":1978075829,"name":"Ipsam quis aut.","trackIDs":[1480478238,1629895974,393565112,1968683977]}],"total":689246006},"required":["total","resources"]},"PlaylistGetAccountPlaylistResponseBody":{"title":"PlaylistGetAccountPlaylistResponseBody","type":"object","properties":{"id":{"type":"integer","description":"Playlist id","example":1386043651,"format":"int32"},"name":{"type":"string","description":"Playlist name","example":"Nemo aperiam ut et."},"trackIDs":{"type":"array","items":{"type":"integer","example":1249627211,"format":"int32"},"description":"Array of TrackIDs","example":[1761482889,1190159261,710083815]}},"example":{"id":597677693,"name":"Veritatis animi amet dolores odio itaque.","trackIDs":[1616165167,1638017335,805741283]},"required":["id","name","trackIDs"]},"PlaylistRemoveTrackFromAccountPlaylistResponseBody":{"title":"PlaylistRemoveTrackFromAccountPlaylistResponseBody","type":"object","properties":{"message":{"type":"string","description":"Operation completion status","example":"Ut eaque accusantium ullam et quod et."}},"example":{"message":"Quia vel."},"required":["message"]},"PlaylistRenameAccountPlaylistRequestBody":{"title":"PlaylistRenameAccountPlaylistRequestBody","type":"object","properties":{"name":{"type":"string","description":"New playlist name","example":"Voluptatem sit."}},"example":{"name":"Minima illum iste iusto aut non."}},"PlaylistRenameAccountPlaylistResponseBody":{"title":"PlaylistRenameAccountPlaylistResponseBody","type":"object","properties":{"message":{"type":"string","description":"Operation completion status","example":"Ea neque consequatur autem optio."}},"example":{"message":"Vero ut optio possimus rem quisquam magnam."},"required":["message"]}}}